+++ tests/Config/Processor/ConfigProcessorTest.php
@@ -1,0 +1,15 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+
+use Peak\Config\Processor\ConfigProcessor;
+
+class ConfigProcessorTest extends TestCase
+{
+    public function testException()
+    {
+        $this->expectException(\Peak\Config\Exception\ProcessorTypeException::class);
+        $processor = new ConfigProcessor();
+        $processor->process([]);
+    }
+}